rename folder and let client handle election_id
