+++ decidim-bulletin_board-ruby/lib/decidim/bulletin_board/authority/get_election_status.rb
@@ -2,7 +2,7 @@
 
 module Decidim
   module BulletinBoard
-    module Election
+    module Authority
       # This command uses the GraphQL client to get the status of the election.
       class GetElectionStatus
         include Wisper::Publisher
